move testing utils into cf-style-provider
