+++ packages/cf-style-provider/package.json
@@ -25,8 +25,10 @@
     "fela-plugin-unit": "^5.0.0",
     "fela-plugin-validator": "^5.0.0",
     "inline-style-prefixer": "^3.0.5",
+    "js-beautify": "^1.6.14",
     "prop-types": "^15.5.8",
-    "react-fela": "^5.0.0"
+    "react-fela": "^5.0.0",
+    "react-test-renderer": "^15.6.1"
   },
   "peerDependencies": {
     "react": "^0.14.9 || ^15.0.0-0 || ^16.0.0-0"
