+++ packages/cf-style-provider/src/felaTestContext.js
@@ -1,0 +1,9 @@
+// this is a wrapper for testing, it passes fela based
+// components fela's renderer and global theme
+
+import React from 'react';
+import StyleProvider from './StyleProvider';
+
+export default (component, renderer) => (
+  <StyleProvider renderer={renderer} dev>{component}</StyleProvider>
+);
