+++ packages/cf-style-provider/test/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { StyleProvider } from '../../cf-style-provider/src/index';
+import { StyleProvider } from '../src/index';
 import { createComponent } from '../../cf-style-container/src/index';
 import { Button } from '../../cf-component-button/src/index';
 import renderer from 'react-test-renderer';
@@ -26,7 +26,7 @@ test('StyleProvider should render styles', () => {
   }));
 
   render(
-    <StyleProvider>
+    <StyleProvider selectorPrefix="cf-">
       <Foo />
     </StyleProvider>,
     htmlNode
